Issue #21 - Refactor Unit Testing, test data updates from latest topic changes.
