+++ .vscode/settings.json
@@ -4,7 +4,7 @@
     "python.testing.unittestArgs": [
       "-v",
       "-s",
-      "./test",
+      "./src",
       "-p",
       "test_*.py"
     ],
