+++ src/routes/test_curriculum_routes.py
